requirements.txt
langgraph>=1.1.3
langchain-openai>=0.3.0
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.1
sentry-sdk>=2.43.0
